auctionBack/requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2019.3.9
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
crypto==1.4.1
cryptography==3.0
daphne==2.5.0
Django==2.2.3
django-cors-headers==2.5.3
django-filter==2.1.0
django-redis==4.12.1
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
msgpack==1.0.0
Naked==0.1.31
pika==1.1.0
Pillow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.10.0
pyOpenSSL==19.1.0
pytz==2019.1
PyYAML==5.1
qcloud-python-sts==3.0.3
redis==3.5.3
requests==2.21.0
selenium==3.141.0
service-identity==18.1.0
shellescape==3.4.1
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.24.3
zope.interface==5.1.0
